+++ app/javascript/controllers/session_controller.js
@@ -2,6 +2,10 @@ import { Controller } from "@hotwired/stimulus"
 
 // Connects to data-controller="session"
 export default class extends Controller {
+  static targets = [
+    "phonediv"
+  ]
+
   connect() {
   }
 
@@ -42,11 +46,11 @@ export default class extends Controller {
         .then(response => response.json())
         .then(data => {
           if (!data.exists) {
-            phoneInput.parentElement.classList.remove("border", "border-green-500");
-            phoneInput.parentElement.classList.add("border", "border-yellow-500");
+            this.phonedivTarget.classList.remove("border-green-500");
+            this.phonedivTarget.classList.add("border", "border-yellow-500");
           } else {
-            phoneInput.parentElement.classList.remove("border", "border-yellow-500");
-            phoneInput.parentElement.classList.add("border", "border-green-500");
+            this.phonedivTarget.classList.remove("border-yellow-500");
+            this.phonedivTarget.classList.add("border-green-500");
             document.querySelector("#name").disabled = true;
             document.querySelector("#email").disabled = true;
             document.getElementById("submit-otp").click();