+++ test/support/unifi_api_stubs.rb
@@ -54,7 +54,7 @@ module UnifiApiStubs
       }.to_json, headers: { "Content-Type" => "application/json" })
   end
 
-  def stub_guest_authorization_api(base_url = "https://heimdall.test", site_id: "88f7af54-98f8-306a-a1c7-c9349722b1f6", client_id: "497f6eca\-6276\-4993\-bfeb\-53cbbbba6f08", api_key: "test-api-key")
+  def stub_unifi_authorize_api(base_url = "https://heimdall.test", site_id: "88f7af54-98f8-306a-a1c7-c9349722b1f6", client_id: "497f6eca\-6276\-4993\-bfeb\-53cbbbba6f08", api_key: "test-api-key")
     stub_request(:post, %r{#{base_url}/proxy/network/integration/v1/sites/#{site_id}/clients/#{client_id}/actions})
       .to_return(status: 200, body: {
         "action": "AUTHORIZE_GUEST_ACCESS",